requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
gdown==4.7.1
tqdm==4.66.1
scikit-learn==1.3.0
matplotlib==3.8.0
Pillow==10.0.0
